Implement Cypress plugin to handle logging in and checking alternative user session (simulating multi session along with the browser existing one); Add E2E test for logging out from all sessions;
